requirements.txt
beautifulsoup4>=4.12.0
numpy>=1.24.0
Pillow>=10.0.0
python-whois>=0.8.0
requests>=2.31.0
fpdf2>=2.7.0
pytest==8.4.2
